requirements.txt
fastapi==0.92.0
pydantic==1.10.5
uvicorn==0.20.0
openai==0.27.6
weaviate-client==3.16.2